Changed the DLM to use sync.pool.
Every time a DLM lock is acquired and released the DLM code allocates and
frees a localLockTrack (unless a second thread gets the lock while the
first goroutine is holding it).  localLockTrack is the actual DLM lock
(rendevouz point) and now includes a trackedlock.RWMutexTrack to add to
its size.

When a tarball containing a some ~72,000 small files and 1.5 Gbyte is
untarred in the FUSE mountpoint of a file system, we see about 1,279,000
Mutexes and 1,109,000 RWMutex-like objects created (probably all of the
RWMutex-like objects are DLM locks).  that's ~17.8 mutex and ~15.4
RWmutex-like objects per file!

atop shows that proxyfsd had grown to a VSIZE of 820.7M and RSIZE
of 428.4M.