Changed swiftclient to use sync.Pool for chunked connections
We create and destroy a lot of chunkedPutContextStruct (at least
one per file copied to ProxyFS) and each one includes a Mutex and a
sortedmap.LLRBtree (which includes another Mutex).

To cut down on the number of objects that are allocated and
discarded (to help out garbage collection), use a sync.Pool to hold
chunkedPutContextStruct so they can be reused.