Trim extra bits from Samba chmod request.
Samba sends the type (mode) of the file as part of a chmod request,
but the recent change to MountHandle.SetStat() returns EINVAL if any
bits past the first 12 bits are set, so Sambas request to set the
permissions to 010444 failed.

Previously MountHandle.SetStat() failed only if bits past the first 32
bits were set. the routine which actually changes the permission bits,
(*volumeStruct).SetPermMode(), calls inode.determineMode() which masked
the new request with 0777 (which also prevents the setuid, setgid,
and sticky bits from being set or queried).

We'll tackle support for the setuid, setgid, and sticky bits later.